requirements.txt
pandas
streamlit
numpy
seaborn
matplotlib.pyplot
plotly.express



